fix: allow root SRV, MX records
